+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 3.2.2
+
+- Supports a feature to disable heading numbering #175
+- Support center image with margin auto #171
+- Support deprecrated align attribute for block #171
+- Fix parsing of style attribute with a key with no value
+- Improve parsing of style attribute to avoid an extra call to HtmlDecode
+- Extend support of nested list for non-W3C compliant html #173
+- Change way to apply table 100% width
+- Allow to apply percentage widths cells
+
 ## 3.2.1
 
 - Fix indentation of numbering list #166